requirement.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
bs4==0.0.1
coverage==5.3.1
distlib==0.3.1
Django==3.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-multiselectfield==0.1.12
et-xmlfile==1.0.1
filelock==3.0.12
jdcal==1.4.1
Js2Py==0.70
mod-wsgi==4.7.1
mysqlclient==2.0.3
openpyxl==3.0.5
Pillow==8.1.0
pyjsparser==2.7.1
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tzlocal==2.1
virtualenv==20.3.1
virtualenvwrapper-win==1.2.6
